position は Int に統一。
MediaPlayer が int だから。
DB は Long のまま。

+++ app/src/main/kotlin/jp/ddo/masm11/contextplayer/player/Player.kt
@@ -338,11 +338,11 @@ class Player : Runnable {
 	}
     }
     
-    val currentPosition : Long
+    val currentPosition: Int
     get() {
 	val plr = curPlayer
 	if (plr != null)
-	    return plr.currentPosition.toLong()
+	    return plr.currentPosition
 	return 0
     }
